Add files via upload
This script reads SEM or TEM image stack and prints out the list of files that are out of focus. TIF format is common for EM images. If you have problem loading the images, check the number of channels in your file (i.e. RGB). You can import files as jpeg too.